--- Assets/B_Main_script.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f494a353f9365a84fa5a192590aa0cb1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
