--- Assets/NextScene.cs
@@ -1,20 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class NextScene : MonoBehaviour
-{
-    // Start is called before the first frame update
-    
-    private void Start()
-    {
-        
-    }
-
- 
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        SceneManager.LoadScene(1);
-    }
-}
