--- Assets/NextScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b638bf2448bcd4f40804657e07ce03c5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
