--- Assets/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 077c93fe0c70ad047a763122ca9c4d37
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
