+++ Assets/Scenes/vertical.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 75acf7242ec8611409b4a503da94b1ee
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
