--- Assets/ToLevel2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 79f103fdf9a26ad44811ac6905c6385e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
