+++ Assets/scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 314d3c49ae31bdb4cb28498ea52111a2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
