+++ Assets/scripts/buttonScripts.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c09cf49a605a3d744ad3366b66edd50a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
