+++ Assets/scripts/button_Main.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e7a337e430aa1ce4fba7bb0a13a5473f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
