+++ Assets/scripts/checkpoint.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class checkpoint : MonoBehaviour
+{
+    // init some modifiable vars in Unity
+    [Tooltip("Specifies whether the player will respawn with their gravity flipped or not.")]
+    public bool negativeGravity = false;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
