+++ Assets/scripts/checkpoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c6b874713e80154992528317078fad9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
