+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,7 +5,7 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/Main_Menu.unity
     guid: dc4b04be756e1b641b370f5d6d8470d6
   - enabled: 1
@@ -17,4 +17,7 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Scenes/GAME1234.unity
     guid: 1f3129fc6c378274a80db18603bfa1c7
+  - enabled: 1
+    path: Assets/Scenes/vertical.unity
+    guid: 75acf7242ec8611409b4a503da94b1ee
   m_configObjects: {}
